Can now view event without logging in Fixed displaying image

+++ app/Http/Controllers/EventsController.php
@@ -9,6 +9,7 @@ use Illuminate\Http\Request;
 use App\EventTag;
 use Illuminate\View\View;
 use function PhpParser\filesInDir;
+use Auth;
 
 class EventsController extends Controller
 {
@@ -128,23 +129,28 @@ class EventsController extends Controller
 
     public function join($id)
     {
-
-        $event = Event::findOrFail($id);
-        if (!$event->participants->contains(auth()->user()->id) && ($event->owner->id != auth()->user()->id)) {
-            $event->participants()->attach(auth()->user()->id);
+        if(Auth::check()) {
+            $event = Event::findOrFail($id);
+            if (!$event->participants->contains(auth()->user()->id) && ($event->owner->id != auth()->user()->id)) {
+                $event->participants()->attach(auth()->user()->id);
+            }
+            //TODO: Add error 'You already joined!'
         }
+        //TODO: Add error 'You are not logged in!'
         return redirect('/events/' . $event->id);
-        //TODO: Add error 'You already joined!'
     }
 
     public function leave($id)
     {
-        $event = Event::findOrFail($id);
-        if ($event->participants->contains(auth()->user()->id) && ($event->owner->id != auth()->user()->id)) {
-            $event->participants()->detach(auth()->user()->id);
+        if(Auth::check()) {
+            $event = Event::findOrFail($id);
+            if ($event->participants->contains(auth()->user()->id) && ($event->owner->id != auth()->user()->id)) {
+                $event->participants()->detach(auth()->user()->id);
+            }
+            //TODO: Add error 'You are not joined!'
         }
+        //TODO: Add error 'You are not logged in!'
         return redirect('/events/' . $id);
-        //TODO: Add error 'You already joined!'
     }
 
     private function formatDate(){
